# Conflicts:
#	.idea/gradle.xml
#	.idea/misc.xml
#	.idea/modules.xml
#	server/src/main/java/com/omc/domain/cancel/entity/Cancel.java
#	server/src/main/java/com/omc/domain/cancel/service/CancelService.java
#	server/src/main/java/com/omc/domain/img/repository/ImgRepository.java
#	server/src/main/java/com/omc/domain/member/api/MemberController.java
#	server/src/main/java/com/omc/domain/member/api/SellerController.java
#	server/src/main/java/com/omc/domain/member/dto/MemberResponseDto.java
#	server/src/main/java/com/omc/domain/member/dto/SignUpRequestDto.java
#	server/src/main/java/com/omc/domain/member/dto/SingleParamDto.java
#	server/src/main/java/com/omc/domain/member/entity/AuthMember.java
#	server/src/main/java/com/omc/domain/member/entity/Member.java
#	server/src/main/java/com/omc/domain/member/repository/RefreshTokenRepository.java
#	server/src/main/java/com/omc/domain/member/service/CustomMemberDetailService.java
#	server/src/main/java/com/omc/domain/member/service/MemberService.java
#	server/src/main/java/com/omc/domain/product/api/ProductController.java
#	server/src/main/java/com/omc/domain/product/dto/ProductDto.java
#	server/src/main/java/com/omc/domain/product/entity/Product.java
#	server/src/main/java/com/omc/domain/product/service/ProductService.java
#	server/src/main/java/com/omc/domain/report/api/ReportController.java
#	server/src/main/java/com/omc/domain/report/dto/ReportDto.java
#	server/src/main/java/com/omc/domain/report/entity/Report.java
#	server/src/main/java/com/omc/domain/report/repository/ReportRepository.java
#	server/src/main/java/com/omc/domain/report/service/ReportService.java
#	server/src/main/java/com/omc/domain/reservation/api/ReservationController.java
#	server/src/main/java/com/omc/domain/reservation/dto/ReservationDto.java
#	server/src/main/java/com/omc/domain/reservation/entity/Reservation.java
#	server/src/main/java/com/omc/domain/reservation/service/ReservationService.java
#	server/src/main/java/com/omc/domain/review/api/ReviewController.java
#	server/src/main/java/com/omc/domain/review/dto/ReviewDto.java
#	server/src/main/java/com/omc/domain/review/entity/Review.java
#	server/src/main/java/com/omc/domain/review/repository/ReviewRepository.java
#	server/src/main/java/com/omc/domain/review/service/ReviewService.java
#	server/src/main/java/com/omc/global/config/MailConfig.java
#	server/src/main/java/com/omc/global/config/security/config/SecurityConfig.java
#	server/src/main/java/com/omc/global/config/security/handler/OAuth2AuthenticationSuccessHandler.java
#	server/src/main/java/com/omc/global/error/ErrorCode.java
#	server/src/main/java/com/omc/global/error/ErrorResponse.java
#	server/src/main/java/com/omc/global/jwt/TokenProvider.java
#	server/src/main/java/com/omc/global/jwt/filter/JwtAuthenticationFilter.java
#	server/src/main/java/com/omc/global/jwt/filter/JwtFilter.java
#	server/src/main/java/com/omc/global/util/Util.java
#	server/src/main/resources/application.yml

+++ server/src/main/java/com/omc/domain/product/service/ProductService.java
@@ -16,11 +16,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.omc.domain.img.dto.ImgDto;
-import com.omc.domain.img.entity.Img;
+import com.omc.domain.img.entity.ProductImg;
 import com.omc.domain.img.repository.ImgRepository;
 import com.omc.domain.member.entity.Member;
 import com.omc.domain.member.entity.UserRole;
-import com.omc.domain.member.repository.MemberRepository;
 import com.omc.domain.product.dto.ProductDto;
 import com.omc.domain.product.dto.StopDto;
 import com.omc.domain.product.entity.Facilities;
@@ -45,7 +44,6 @@ import lombok.extern.slf4j.Slf4j;
 @Service
 public class ProductService {
 
-	private final MemberRepository memberRepository;
 	private final ProductRepository productRepository;
 	private final FacilitiesRepository facilitiesRepository;
 	private final LocationRepository locationRepository;
@@ -65,17 +63,18 @@ public class ProductService {
 	@SneakyThrows
 	@Transactional
 	public void create(ProductDto.Request req,
-					   List<MultipartFile> multipartFiles) {
+					   List<MultipartFile> multipartFiles,
+					   Member member) {
 
 		log.info("multipartFile imgs={}", multipartFiles);
 		String address = req.getAddress();
 
-		List<Img> imgList = uploadImgAndImgDtoToEntity(multipartFiles);
+		List<ProductImg> productImgList = uploadImgAndImgDtoToEntity(multipartFiles);
 
 		Product product = Product.builder()
 								 .subject(req.getSubject())
 								 .description(req.getDescription())
-								 .imgList(imgList)
+								 .productImgList(productImgList)
 								 .address(address)
 								 .zipcode(req.getZipcode())
 								 .price(req.getPrice())
@@ -86,6 +85,7 @@ public class ProductService {
 								 .star(0.0)
 								 .reportCount(0L)
 								 .count(0L)
+								 .member(member)
 								 .build();
 
 		saveLocation(product, req);
@@ -105,9 +105,16 @@ public class ProductService {
 	@Transactional
 	public void update(ProductDto.Request req,
 					   List<MultipartFile> multipartFiles,
-					   Long productId) {
+					   Long productId,
+					   Member member) {
 
 		Product findProduct = ifExistReturnProduct(productId);
+
+		if (!findProduct.getMember().getId().equals(member.getId()) || !member.getUserRole()
+																			  .equals(UserRole.ROLE_ADMIN)) {
+			throw new BusinessException(ErrorCode.NOT_PRODUCT_WRITER);
+		}
+
 		findProduct.editProduct(req);
 
 		if (req.getFacilities() != null) {
@@ -120,10 +127,10 @@ public class ProductService {
 		}
 
 		if (multipartFiles != null) {
-			findProduct.getImgList().stream().map(Img::getImgName).forEach(s3Service::deleteFile);
+			findProduct.getProductImgList().stream().map(ProductImg::getImgName).forEach(s3Service::deleteFile);
 			imgRepository.deleteByProductId(productId);
-			List<Img> imgs = uploadImgAndImgDtoToEntity(multipartFiles);
-			findProduct.setImgList(imgs);
+			List<ProductImg> productImgs = uploadImgAndImgDtoToEntity(multipartFiles);
+			findProduct.setProductImgList(productImgs);
 		}
 
 		productRepository.save(findProduct);
@@ -136,11 +143,21 @@ public class ProductService {
 	 * @return 상품 정보
 	 */
 	@Transactional
-	public ProductDto.Response getProduct(Long productId) {
+	public ProductDto.Response getProduct(Long productId, Member member) {
 		Product findProduct = productRepository.findById(productId)
 											   .orElseThrow(() -> new BusinessException(ErrorCode.PRODUCT_NOT_FOUND));
 
+		Optional<LikeHistory> likeHistory = Optional.empty();
+
+		if (member != null) {
+			likeHistory = likeHistoryRepository.findByMemberIdAndProductId(member.getId(),
+																		   findProduct.getId());
+		}
+
+		findProduct.addViews();
+
 		return ProductDto.Response.builder()
+								  .id(findProduct.getId())
 								  .subject(findProduct.getSubject())
 								  .description(findProduct.getDescription())
 								  .address(findProduct.getAddress())
@@ -155,7 +172,7 @@ public class ProductService {
 								  .checkOut(findProduct.getCheckOut())
 								  .img(getImgs(productId))
 								  .likes(findProduct.getLikes())
-								  // .isLike() todo : Member 적용 후 수정, 추천 여부, 회원일 경우 추천 여부 확인해서 넣어줘야함
+								  .isLike(likeHistory.isPresent())
 								  .build();
 	}
 
@@ -170,9 +187,10 @@ public class ProductService {
 		// todo 리팩터링
 
 		String sortBy = switch (search.getSort()) { // 최신순, 인기순, 조회순, 추천순
-			case "추천" -> "likes";
-			case "조회" -> "views";
-			case "인기" -> "reservations";
+			case "0" -> "id"; // 최신순
+			case "1" -> "star"; // 인기순
+			case "2" -> "views"; // 조회순
+			case "3" -> "likes"; // 추천순
 			default -> "id";
 		};
 
@@ -236,20 +254,20 @@ public class ProductService {
 	public Page<Product> getMyProductList(Member member, ProductDto.Search search) {
 
 		if (member.getUserRole() == UserRole.ROLE_USER) {
-			throw new BusinessException(ErrorCode.TEST); // todo : member 적용 후 에러코드 수정
+			throw new BusinessException(ErrorCode.NOT_PRODUCT_WRITER);
 		}
 
 		String sortBy = switch (search.getSort()) { // 최신순, 인기순, 조회순, 추천순
-			case "추천" -> "likes";
-			case "조회" -> "views";
-			case "인기" -> "reservations";
+			case "0" -> "id"; // 최신순
+			case "1" -> "star"; // 인기순
+			case "2" -> "views"; // 조회순
+			case "3" -> "likes"; // 추천순
 			default -> "id";
 		};
 
 		Pageable pageable = PageRequest.of(Math.toIntExact(search.getPage() - 1),
 										   Math.toIntExact(search.getSize()),
 										   Sort.by(sortBy).descending());
-		// return productRepository.findAllBySellerId(member.getId(), pageable);
 		return productRepository.findAllByMemberId(member.getId(), pageable);
 	}
 
@@ -259,9 +277,16 @@ public class ProductService {
 	 * @param productId : 상품 id
 	 */
 	@Transactional
-	public void delete(Long productId) {
+	public void delete(Long productId, Member member) {
 		Product findProduct = ifExistReturnProduct(productId);
-		findProduct.getImgList().stream().map(Img::getImgName).forEach(s3Service::deleteFile);
+		findProduct.getProductImgList().stream().map(ProductImg::getImgName).forEach(s3Service::deleteFile);
+
+		if (!findProduct.getMember().getId().equals(member.getId()) || !member.getUserRole()
+																			  .equals(UserRole.ROLE_ADMIN)) {
+			throw new BusinessException(ErrorCode.NOT_PRODUCT_WRITER);
+		}
+
+		findProduct.getProductImgList().stream().map(ProductImg::getImgName).forEach(s3Service::deleteFile);
 		productRepository.delete(findProduct);
 	}
 
@@ -274,7 +299,7 @@ public class ProductService {
 	@Transactional
 	public void likeProduct(Long productId, Member member) {
 		Product product = ifExistReturnProduct(productId);
-		// todo 존재하는 회원인지, 일반 유저인지 확인하는 메서드 추가
+
 		Optional<LikeHistory> likeHistory = likeHistoryRepository.findByMemberIdAndProductId(member.getId(),
 																							 product.getId());
 		if (likeHistory.isPresent()) {
@@ -305,11 +330,11 @@ public class ProductService {
 		UserRole userRole = member.getUserRole();
 
 		if (!product.getMember().getId().equals(member.getId()) || userRole != UserRole.ROLE_ADMIN) {
-			throw new BusinessException(ErrorCode.TEST); // todo member 추가 후 수정
+			throw new BusinessException(ErrorCode.NOT_PRODUCT_WRITER);
 		}
 
 		if (product.getIsStop() == 2) {
-			throw new BusinessException(ErrorCode.TEST); // todo 관리자만 블라인드 해제 가능
+			throw new BusinessException(ErrorCode.FORBIDDEN_ADMIN);
 		}
 
 		ArrayList<StopHistory> stopHistories = new ArrayList<>();
@@ -346,8 +371,8 @@ public class ProductService {
 	 */
 	private List<String> getImgs(Long productId) {
 		return imgRepository.findAllByProductId(productId).stream()
-							.map(Img::getImgUrl)
-							.map(s -> "value\":\"" + s)
+							.map(ProductImg::getImgUrl)
+							// .map(s -> "value\":\"" + s)
 							.toList();
 	}
 
@@ -360,7 +385,7 @@ public class ProductService {
 	private List<String> getFacilities(Long productId) {
 		return facilitiesRepository.findAllByProductId(productId).stream()
 								   .map(Facilities::getKeyword)
-								   .map(s -> "value\":\"" + s)
+								   // .map(s -> "value\":\"" + s)
 								   .toList();
 	}
 
@@ -373,23 +398,23 @@ public class ProductService {
 	private List<String> getLocations(Long productId) {
 		return locationRepository.findAllByProductId(productId).stream()
 								 .map(Location::getKeyword)
-								 .map(s -> "value\":\"" + s)
+								 // .map(s -> "value\":\"" + s)
 								 .toList();
 	}
 
 	/**
 	 * ImgDto를 Img로 변환
 	 *
-	 * @param imgList    : Img 리스트
-	 * @param imgDtoList : ImgDto 리스트
+	 * @param productImgList : ProductImg 리스트
+	 * @param imgDtoList     : ImgDto 리스트
 	 */
-	private static void imgDtoToImg(List<Img> imgList, List<ImgDto.Request> imgDtoList) {
+	private static void imgDtoToImg(List<ProductImg> productImgList, List<ImgDto.Request> imgDtoList) {
 		for (ImgDto.Request imgDto : imgDtoList) {
-			Img img = Img.builder()
-						 .imgName(imgDto.getImgName())
-						 .imgUrl(imgDto.getImgUrl())
-						 .build();
-			imgList.add(img);
+			ProductImg productImg = ProductImg.builder()
+											  .imgName(imgDto.getImgName())
+											  .imgUrl(imgDto.getImgUrl())
+											  .build();
+			productImgList.add(productImg);
 		}
 	}
 
@@ -399,17 +424,17 @@ public class ProductService {
 	 * @param multipartFiles : 상품 이미지
 	 * @return : 이미지 리스트
 	 */
-	private List<Img> uploadImgAndImgDtoToEntity(List<MultipartFile> multipartFiles) {
-		List<Img> imgList = new ArrayList<>();
+	private List<ProductImg> uploadImgAndImgDtoToEntity(List<MultipartFile> multipartFiles) {
+		List<ProductImg> productImgList = new ArrayList<>();
 		List<ImgDto.Request> imgDtoList = new ArrayList<>();
 
 		for (MultipartFile img : multipartFiles) {
 			ImgDto.Request imgDto = s3Service.uploadImage(img, "product");
 			imgDtoList.add(imgDto);
 		}
-		imgDtoToImg(imgList, imgDtoList);
+		imgDtoToImg(productImgList, imgDtoList);
 
-		return imgList;
+		return productImgList;
 	}
 
 	/**
@@ -476,7 +501,7 @@ public class ProductService {
 	 * @param productId : 상품 id
 	 * @return : 상품
 	 */
-	private Product ifExistReturnProduct(Long productId) {
+	public Product ifExistReturnProduct(Long productId) {
 		return productRepository.findById(productId)
 								.orElseThrow(() -> new BusinessException(ErrorCode.PRODUCT_NOT_FOUND));
 	}
@@ -488,7 +513,9 @@ public class ProductService {
 	 * @return : 상품 목록 (ResponseDto)
 	 */
 	public List<ProductDto.Response> convertToResponse(List<Product> content) {
+
 		List<ProductDto.Response> responseList = new ArrayList<>();
+
 		for (Product product : content) {
 			responseList.add(ProductDto.Response.builder()
 												.id(product.getId())
